src/api/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
databases==0.9.0
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.2
fonttools==4.53.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-dialogflow==2.30.0
googleapis-common-protos==1.63.0
graphene==3.3
graphene-sqlalchemy==2.1.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_metadata==7.2.0
importlib_resources==6.4.0
kiwisolver==1.4.5
Markdown==3.6
matplotlib==3.7.5
numpy==1.24.4
openai==1.34.0
packaging==24.0
pandas==2.0.3
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
singledispatch==4.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.30
starlette==0.36.3
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zipp==3.19.1
